inventory-api/requirements.txt
fastapi
uvicorn
pydantic
SQLAlchemy
pg8000
cloud-sql-python-connector[pg8000]
google-cloud-secret-manager
google-cloud-logging
